Refactor some of the utility methods to CosmosDBUtil class
